split slack module
